Test file
This file (tester) works perfectly, and the original (contactUs) is not working as it should. The files have the same code and should give the same layout/output. However, when I run the tester I have the page responsive as it should be, while on the original file, I still have the same error as before (image overlaps the form). Could you guys test it too? Then, we can delete the wrong one.